# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Interactive Documentation for Elyte-Realm API",
        "title": "Elyte Realm API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Elyte Fiber Application.",
            "url": "https://github.com/elyte5star.",
            "email": "elyte5star@gmail.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/auth/form-login": {
            "post": {
                "description": "Create a new bearer token.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Create a new bearer token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username/Email",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RequestResponse"
                        }
                    },
                    "400": {
                        "description": "BAD REQUEST",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "423": {
                        "description": "LOCKED",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "503": {
                        "description": "SERVICE UNAVAILABLE",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "description": "Create a new bearer token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Create a new bearer token",
                "parameters": [
                    {
                        "description": "Login data",
                        "name": "credential",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RequestResponse"
                        }
                    },
                    "400": {
                        "description": "BAD REQUEST",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "423": {
                        "description": "LOCKED",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "503": {
                        "description": "SERVICE UNAVAILABLE",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.LoginRequest": {
            "type": "object",
            "required": [
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 5
                },
                "username": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5
                }
            }
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean",
                    "default": false
                }
            }
        },
        "response.RequestResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "result": {},
                "success": {
                    "type": "boolean",
                    "default": true
                },
                "timeStamp": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer Token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}